pyproject.toml
[project]
name = "multimodal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "flask>=3.1.3",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]
